Temporarily hide organisations
This was confusing my state hosts, who believed the portal to be the source of truth about sponsors. 
In reality, yes, but it's also massively out of date. So I have commented it out for now.
When we get around to dev on sponsors, please create a new branch then merge back into master when complete.